requirements.txt
airflow_provider_great_expectations==0.2.6
apache_flink==1.17.1
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.6.1
apache-airflow-providers-docker==3.7.5
avro_python3==1.9.2.1
confluent_kafka==2.2.0
pandas
pendulum==3.0.0
psycopg2_binary==2.9.7
pyarrow==8.0.0
pydeequ==1.2.0
pymongo==3.13.0
python-dotenv
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy
numpy